requirements.txt
asgiref==3.11.1
Django==6.0.3
django-taggit==5.0.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.2.0
Markdown==3.10.2
packaging==26.0
psycopg2-binary==2.9.11
PyJWT==2.12.1
python-decouple==3.8
sqlparse==0.5.5
tzdata==2025.3
